refactor(message): separate message to constants
