+++ src/App.js
@@ -1,6 +1,6 @@
 const User = require('./User');
-const MESSAGE = require('./MESSAGE');
-const { Console, Random } = require('@woowacourse/mission-utils');
+const MESSAGE = require('./constants/message');
+const { Console } = require('@woowacourse/mission-utils');
 const { checkValidLotto, checkValidBonus } = require('./utils/validator');
 
 class App {
